Implemented the 'bonus' API for fetching results
And since it's parameters are actual formatted Dates, I had to roll out a multiplatform
Date implementation as well... which isn't really implemented at all for iOS and Desktop,
but the foundation is there.

For Android, it just uses the kotlinx-datetime LocalDate, and since the formatter requires
API 26+, I say F it and just rolled one on my own. Should do the trick. Too lazy to bother
confirming it with tests now though.