Networking (#2)
* Add initial support for networking via Ktor

And also added it to the DI modules block.
To be expanded on later when it comes to using it.

* Implement the two API calls

And implement the 'fetch 20' POJO data class.

* Implemented the 'bonus' API for fetching results

And since it's parameters are actual formatted Dates, I had to roll out a multiplatform
Date implementation as well... which isn't really implemented at all for iOS and Desktop,
but the foundation is there.

For Android, it just uses the kotlinx-datetime LocalDate, and since the formatter requires
API 26+, I say F it and just rolled one on my own. Should do the trick. Too lazy to bother
confirming it with tests now though.

* Fix the typo in 'fetchResults' url - removed extra space

* Date-related refactors, also included a factory expect/actual method

Well, no point in dates being in the koinModules package; and platform-specific
classes have the 'actual' createDate method implementations too

* Moved the 'status' out of the Fetch20Result constructor

And made the 'stringStatus' private since that is just used to calculate
the actual enum-based status