Multiple functional changes across the board
ActivityMain now initializes Koin in it's onCreate() due to no Application class.
TableBetsRepository and TableBetsRepositoryImpl are added to the Koin module.
FragmentBettingTable now pulls in the repository via DI and uses it when selecting items.

However, it's still using a temporary ID. This needs to be passed in from the main screen.
So that needs to be the next change...