Added permission, moved koin initialization
Added internet permission to AndroidManifest.xml
while i'm there, might as well make an Application class and declare it

So, also added a barebones MyApplication class, and moved koin initialziation from
ActivityMain to MyApplication. And things seem to work out fine.

But the app is crashing with
Caused by: kotlinx.serialization.SerializationException: Serializer for class 'Fetch20Result' is not found.

so more things to take care of...