Add all necessary junk to Fetch20Result
I wanted to skip deserializing this nonsense and just get them as strings,
but noo, this dumbass framework can't grab things as literal strings or json objects
and requires me to make full deserialization for this... so here it is.

The object deserializes fine, the UI starts up fine, so i can get back to more UI work now.