UI implementation (#3)
* Add support for jetpack navigation

Add compose navigation this time actually.

* ActivityMain refactor, starting to integrate FragmentGameList

* Initial FragmentGameList implementation

* Initial FragmentBettingTable implementation

Ok, so they look like crap, and are all lined up vertically instead of forming
a MxN matrix, but there's 80 of them here. Clicking on the number really changes
the background.

BUGS TO FIX:
- number alignment
- get rid of the purple button background overlaying the circle which *should* be the only background

* Initial FragmentDrawResults implementation and hooked up everything to ActivityMain

Well... The FragmentDrawResults has the identical 'vertical' alignment problem like
the betting table one.

It's also missing the two repositories and the API calls as well to make it functional,
but it's a placeholder i suppose.

BUGS TO FIX:
- the 'table' should look similar to the betting table view
- it should also list all played tables instead of using random junk data

* Fixed issues on FragmentBettingTable

Ok, now the numbers are centered in their circles, using gray background
if they aren't selected, green background if they are selected.

Time to fix the FragmentDrawResults

* Multiple functional changes across the board

ActivityMain now initializes Koin in it's onCreate() due to no Application class.
TableBetsRepository and TableBetsRepositoryImpl are added to the Koin module.
FragmentBettingTable now pulls in the repository via DI and uses it when selecting items.

However, it's still using a temporary ID. This needs to be passed in from the main screen.
So that needs to be the next change...

* More necessary 'framework' stuff

- Made AvailableGamesRepository and it's implementation
- added it to the koin common module
- MainActivity pulls in the AvailableGamesRepository after initializing koin and forces a network fetch
- FragmentGameList will use the fetched results to populate it's list of items

crashed due to missing internet permission...

* Added permission, moved koin initialization

Added internet permission to AndroidManifest.xml
while i'm there, might as well make an Application class and declare it

So, also added a barebones MyApplication class, and moved koin initialziation from
ActivityMain to MyApplication. And things seem to work out fine.

But the app is crashing with
Caused by: kotlinx.serialization.SerializationException: Serializer for class 'Fetch20Result' is not found.

so more things to take care of...

* Fix kotlinx serialization plugin issue

Turns out, i didn't even apply the plugin, so the whole @Serializable annotation
did nothing, and that's why the .serializer() method wasn't being generated.

* Consolidate serialization problems and make repo use apiclient's json declaration

The ApiClient's json client configuration contains all the necessary custom serializers, and
the AvailableGamesRepositoryImpl is now using it.

* Add all necessary junk to Fetch20Result

I wanted to skip deserializing this nonsense and just get them as strings,
but noo, this dumbass framework can't grab things as literal strings or json objects
and requires me to make full deserialization for this... so here it is.

The object deserializes fine, the UI starts up fine, so i can get back to more UI work now.

* Add row above the list of games for FragmentGameList

* Fix app title

* Format game start time properly

* Fixed formatting until draw, the GameList fragment looks like it should be

The only thing missing here is to launch fragment2 with the boardID when an item is clicked.

* Extract navigation from ActivityMain into a navgraph and reuse it in activity

The initial navgraph (navhost) was moved from ActivityMain to a separate file
containing the NavigationGraph, and it's navhost was reused instead of the old
one that was in ActivityMain

Also do some minor cleaning up.

* Address CR comments

* fix the seconds formatting to pull the sign out before the whole time

Because 00:0-5 and 00:-19 looked really dumb. So make it look like -00:19 instead.