FragmentWebView and it's actual Compose equivallent...
So making the webview fragment was the typical thing to do, except
normal fragments can be used at all within the Compose framework

So a different approach was needed and that's what WebViewScreen is for.
I'm just kinda sad to delete the last 30+mins of work so i'm gonna commit it
then delete it in follow up commit...

The centering of the webview component is kinda crap but the man said to
not look too much into it so eh, lets just leave it like that.