Implement/fix the getResults() API call - still to be finalized
Ok, at least the API is returning some json to me now... the actual objects,
serialization, and finalizing is to be done tomorrow.

But tl;dr the API can't return results between yesterday and tomorrow
(both dates are somehow out of range), but from today to today seems to work fine.

So, that's something at least.

Tomorrow's workload:
- fix the FragmentDrawResults's UI because this one is whack and looks nothing like what I want
- parse teh API response instead of just returning it as a string
- save what needs to be saved in the TableBetsRepository
- whatever else comes up...